requirements.txt
Flask>=3.1.1
Pillow>=11.3.0
opencv-python>=4.12.0
ffmpeg-python>=0.2.0
Werkzeug>=3.1.3
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
